refactor(manager/pixi): extract shared pixi.lock update helper
